requirements.txt
Django==3.0.5
psycopg2-binary==2.8.6
django-modeltranslation==0.17.3
gunicorn==20.1.0
google-cloud-secret-manager==2.12.6
google-cloud-storage==2.7.0